msm: Update Ion and IOMMU APIs to use dma_addr_t
The current APIs for Ion and the custom IOMMU mapping assume that
many parameters are unsigned long. This type is not big enough to
hold physical addresses on LPAE systems. Update the APIs to accomodate
these types. Because we are updating the APIs, all clients need to be
updated as well, either by using the correct type or using a cast
where necessary.

+++ drivers/char/adsprpc.c
@@ -630,7 +630,7 @@ static int get_args(uint32_t kernel, struct smq_invoke_ctx *ctx,
 		}
 		list[i].num = 1;
 		pages[list[i].pgidx].addr =
-			buf_page_start((void *)(pbuf->phys +
+			buf_page_start((void *)(unsigned long)(pbuf->phys +
 						 (pbuf->size - rlen)));
 		pages[list[i].pgidx].size =
 			buf_page_size(pra[i].buf.len);
